test code cleanup
