+++ liquibase-integration-tests/src/test/resources/liquibase/extension/testing/command/changelogSync.test.groovy
@@ -1,15 +1,11 @@
 package liquibase.extension.testing.command
 
-import liquibase.Scope
-import liquibase.changelog.ChangeLogHistoryServiceFactory
-import liquibase.changelog.RanChangeSet
-import liquibase.database.Database
+
 import liquibase.change.ColumnConfig
 import liquibase.change.core.CreateTableChange
 import liquibase.change.core.TagDatabaseChange
 import liquibase.exception.CommandValidationException
-
-import static org.junit.Assert.assertNotNull
+import liquibase.util.TestUtil
 
 CommandTests.define {
     command = ["changelogSync"]
@@ -78,12 +74,7 @@ Optional Args:
         }
 
         expectations = {
-            def database = (Database) Scope.getCurrentScope().get("database", null)
-            def changelogHistoryService = Scope.getCurrentScope().getSingleton(ChangeLogHistoryServiceFactory.class).getChangeLogService(database)
-            List<RanChangeSet> ranChangeSets = changelogHistoryService.getRanChangeSets()
-            for (RanChangeSet ranChangeSet : ranChangeSets) {
-                assertNotNull(ranChangeSet.getDeploymentId())
-            }
+            TestUtil.assertAllDeploymentIdsNonNull()
         }
     }
