+++ liquibase-integration-tests/src/test/resources/liquibase/extension/testing/command/updateCount.test.groovy
@@ -1,12 +1,8 @@
 package liquibase.extension.testing.command
 
-import liquibase.Scope
-import liquibase.changelog.ChangeLogHistoryServiceFactory
-import liquibase.changelog.RanChangeSet
-import liquibase.database.Database
-import liquibase.exception.CommandValidationException
 
-import static org.junit.Assert.assertNotNull
+import liquibase.exception.CommandValidationException
+import liquibase.util.TestUtil
 
 CommandTests.define {
     command = ["updateCount"]
@@ -62,12 +58,7 @@ Optional Args:
         ]
 
         expectations = {
-            def database = (Database) Scope.getCurrentScope().get("database", null)
-            def changelogHistoryService = Scope.getCurrentScope().getSingleton(ChangeLogHistoryServiceFactory.class).getChangeLogService(database)
-            List<RanChangeSet> ranChangeSets = changelogHistoryService.getRanChangeSets()
-            for (RanChangeSet ranChangeSet : ranChangeSets) {
-                assertNotNull(ranChangeSet.getDeploymentId())
-            }
+            TestUtil.assertAllDeploymentIdsNonNull()
         }
     }
